Partial test restructure CONT
